@@ -1,45 +1,49 @@
 {
-	"description": "NativeScript Application",
-	"license": "SEE LICENSE IN <your-license-filename>",
-	"readme": "NativeScript Application",
-	"repository": "<fill-your-repository-here>",
-	"nativescript": {
-		"id": "org.nativescript.AngularNativeSeed",
-		"tns-ios": {
-			"version": "3.4.0"
-		},
-		"tns-android": {
-			"version": "3.4.1"
-		}
-	},
-	"dependencies": {
-		"@angular/animations": "~5.1.0",
-		"@angular/common": "~5.1.0",
-		"@angular/compiler": "~5.1.0",
-		"@angular/core": "~5.1.0",
-		"@angular/forms": "~5.1.0",
-		"@angular/http": "~5.1.0",
-		"@angular/platform-browser": "~5.1.0",
-		"@angular/platform-browser-dynamic": "~5.1.0",
-		"@angular/router": "~5.1.0",
-		"@ngrx/store": "^4.1.1",
-		"@ngrx/store-devtools": "^4.1.1",
-		"allow-publish": "^1.0.4",
-		"nativescript-angular": "^5.1.0",
-		"nativescript-checkbox": "^3.0.3",
-		"nativescript-theme-core": "~1.0.4",
-		"ngrx-store-freeze": "^0.2.0",
-		"reflect-metadata": "~0.1.8",
-		"rxjs": "~5.5.2",
-		"tns-core-modules": "~3.4.0",
-		"zone.js": "~0.8.2"
-	},
-	"devDependencies": {
-		"babel-traverse": "6.26.0",
-		"babel-types": "6.26.0",
-		"babylon": "6.18.0",
-		"lazy": "1.0.11",
-		"nativescript-dev-typescript": "~0.6.0",
-		"typescript": "~2.4.2"
-	}
+  "description": "NativeScript Application",
+  "license": "SEE LICENSE IN <your-license-filename>",
+  "readme": "NativeScript Application",
+  "repository": "<fill-your-repository-here>",
+  "nativescript": {
+    "id": "org.nativescript.AngularNativeSeed",
+    "tns-android": {
+      "version": "3.4.1"
+    },
+    "tns-ios": {
+      "version": "3.4.1"
+    }
+  },
+  "dependencies": {
+    "@angular/animations": "~5.2.0",
+    "@angular/cdk": "^5.1.0",
+    "@angular/common": "~5.2.0",
+    "@angular/compiler": "~5.2.0",
+    "@angular/core": "~5.2.0",
+    "@angular/forms": "~5.2.0",
+    "@angular/http": "~5.2.0",
+    "@angular/material": "^5.1.0",
+    "@angular/platform-browser": "~5.2.0",
+    "@angular/platform-browser-dynamic": "~5.2.0",
+    "@angular/router": "~5.2.0",
+    "@ngrx/store": "^4.1.1",
+    "@ngrx/store-devtools": "^4.1.1",
+    "allow-publish": "^1.0.4",
+    "nativescript-angular": "^5.2.0",
+    "nativescript-checkbox": "^3.0.3",
+    "nativescript-material-icons": "^1.0.3",
+    "nativescript-theme-core": "~1.0.4",
+    "ngrx-store-freeze": "^0.2.0",
+    "reflect-metadata": "~0.1.8",
+    "rxjs": "~5.5.2",
+    "tns-core-modules": "~3.4.0",
+    "zone.js": "~0.8.2"
+  },
+  "devDependencies": {
+    "babel-traverse": "6.26.0",
+    "babel-types": "6.26.0",
+    "babylon": "6.18.0",
+    "lazy": "1.0.11",
+    "nativescript-dev-sass": "^1.3.5",
+    "nativescript-dev-typescript": "~0.6.0",
+    "typescript": "^2.6.2"
+  }
 }
