@@ -1,14 +1,22 @@
 import { NgModule, NO_ERRORS_SCHEMA } from '@angular/core';
+import { StoreModule } from '@ngrx/store';
 import { NativeScriptModule } from 'nativescript-angular/nativescript.module';
 import { TNSCheckBoxModule } from 'nativescript-checkbox/angular';
+import { NativeScriptCommonModule } from 'nativescript-angular/common';
 
-import { AppModule } from './app.module';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 
 @NgModule({
   bootstrap: [AppComponent],
-  imports: [AppModule, NativeScriptModule, TNSCheckBoxModule],
+  declarations: [AppComponent],
+  imports: [
+    StoreModule.forRoot({}),
+    AppRoutingModule,
+    NativeScriptModule,
+    NativeScriptCommonModule,
+    TNSCheckBoxModule
+  ],
   schemas: [NO_ERRORS_SCHEMA]
 })
 export class AppNativeModule {}
